Split integer-specific code in core into its own package
Closes #48
